Move code to appropriate location
