+++ src/error_handling/__init__.py
@@ -1,1 +1,1 @@
-from error_handling.error_handling import as_http_exception  # noqa:F401
+from error_handling.error_handling import as_http_exception, http_exception_handler  # noqa:F401
