+++ src/error_handling/error_handling.py
@@ -1,5 +1,12 @@
+import json
+import logging
 import traceback
+import uuid
+from http import HTTPStatus
+
 from fastapi import HTTPException, status
+from pydantic import BaseModel
+from starlette.responses import JSONResponse
 
 
 def as_http_exception(exception: Exception) -> HTTPException:
@@ -13,3 +20,30 @@ def as_http_exception(exception: Exception) -> HTTPException:
             f"{exception}"
         ),
     )
+
+
+class ErrorSchema(BaseModel):
+    detail: str
+    reference: str
+
+
+async def http_exception_handler(request, exc):
+    reference = uuid.uuid4().hex
+    error = ErrorSchema(detail=exc.detail, reference=reference)
+    content = error.dict()
+
+    body = await request.body()
+    log_level = logging.DEBUG
+    if exc.status_code == HTTPStatus.INTERNAL_SERVER_ERROR:
+        log_level = logging.WARNING
+    log_message = str(
+        dict(
+            reference=reference,
+            exception=f"{str(exc)!r}",
+            method=request.scope["method"],
+            path=request.scope["path"],
+            body=json.dumps(json.loads(body)),
+        )
+    )
+    logging.log(log_level, log_message)
+    return JSONResponse(content, status_code=exc.status_code)
